+++ src/server/settings/settings-router.ts
@@ -1,9 +1,11 @@
 import { router } from "../trpc/trpc"
-import { get, save } from "./settings-procedures"
+import { get, save, voiceboxStart, voiceboxStatus } from "./settings-procedures"
 
 export let settings = router({
     settings: {
         get,
-        save
+        save,
+        voiceboxStatus,
+        voiceboxStart
     }
 })